Here is a concise commit message summarizing the changes:
```text
refactor: 리뷰 생성 API 수정 (marketId PathVariable 적용 및 ID 타입 Long 변경)
```

**Breakdown of the changes covered:**
* **Controller**: Updated the review creation endpoint to accept `marketId` as a `@PathVariable` (`/v1/create/{marketId}`).
* **DTO**: Removed `marketId` and `regionId` from `CreateReviewReq` since `marketId` is now handled via the URL path.
* **Entity**: Changed the data types of `marketId` and `regionId` in the `Review` entity from `Integer` to `Long`.
* **Service & Converter**: Updated methods to pass `marketId` from the controller down to the entity builder.

+++ kaka/src/main/java/com/umc/umc10th/kaka/domain/review/entity/Review.java
@@ -19,10 +19,10 @@ public class Review {
     private Long id;
 
     @Column(name = "market_id")
-    private Integer marketId;
+    private Long marketId;
 
     @Column(name = "region_id")
-    private Integer regionId;
+    private Long regionId;
 
     @Column(name = "stars")
     private Float stars;
